backend/requirements.txt
flask
gunicorn
python-dotenv
bcrypt
pyjwt
libsql_client
binance-connector
